add Sport,Device,user,athlete,report

+++ src/main/java/com/mapps/persistence/impl/DeviceDAOImpl.java
@@ -30,7 +30,9 @@ public class DeviceDAOImpl implements DeviceDAO{
     public void deleteDevice(Long deviceId) throws DeviceNotFoundException {
         logger.info("A Device was removed from the database");
         Device deviceAux=getDeviceById(deviceId);
-        entityManager.remove(deviceAux);
+        if(deviceAux!=null){
+            entityManager.remove(deviceAux);
+        }
     }
 
     @Override
